add initial support for a zmq (Pixelator) specific device configurator
